Extend query performance from PR #5573 to the profiles page. See that PR for more details.

Scenario: View profile for user with 107 registrations, 52,289 versions
- Before: 1.8m, 2.0m, 2.0m => AVG=116s
- After: 4.39s, 4.32s, 4.36s => AVG=4.36s (-96%)

*This is after dropping outliers. Saw up to 6.4m for before on cold load.

Another optimization we should do is only select the current page... opened #5658 to track this.